Added RangeValidator, CMake files, basic directory structure, and fixed ValidatorInterface to refer to OptionInterface.

+++ src/lib/include/OptionInterface.h
@@ -1,6 +1,9 @@
 #ifndef ETW_CONFIG_OPTION_INTERFACE_H_
 #define ETW_CONFIG_OPTION_INTERFACE_H_
 
+#include <string>
+#include <stdint.h>
+
 namespace etw {
 
 class OptionInterface {
@@ -8,8 +11,8 @@ class OptionInterface {
 public:
 	virtual int getStringValue(std::string *out) = 0;	
 	virtual int setStringValue(std::string in) = 0;
-	virtual int getNumericValue(int *out) = 0;
-	virtual int setNumericValue(int in) = 0;
+	virtual int getNumericValue(int64_t *out) = 0;
+	virtual int setNumericValue(int64_t in) = 0;
 
 };
